backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
asyncpg==0.29.0
pydantic==2.6.0
pydantic-settings==2.1.0
google-cloud-vision==3.5.0
google-generativeai==0.3.2
python-multipart==0.0.7
python-dotenv==1.0.1
